remhi-simulations/requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
pyyaml>=6.0
pandas>=2.0.0
